style: domain 폴더 내에서 pay 도메인 패키지 별도 분리
