+++ src/main/java/com/dangdang/server/controller/ConnectionAccountDatabaseController.java
@@ -1,7 +1,7 @@
 package com.dangdang.server.controller;
 
-import com.dangdang.server.domain.connectionAccount.application.ConnectionAccountDatabaseService;
-import com.dangdang.server.domain.connectionAccount.dto.AddConnectionAccountRequest;
+import com.dangdang.server.domain.pay.connectionAccount.application.ConnectionAccountDatabaseService;
+import com.dangdang.server.domain.pay.connectionAccount.dto.AddConnectionAccountRequest;
 import com.dangdang.server.domain.member.domain.entity.Member;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
