+++ src/main/java/com/dangdang/server/domain/pay/bankAccount/domain/BankAccountRepository.java
@@ -1,6 +1,6 @@
-package com.dangdang.server.domain.bankAccount.domain;
+package com.dangdang.server.domain.pay.bankAccount.domain;
 
-import com.dangdang.server.domain.bankAccount.domain.entity.BankAccount;
+import com.dangdang.server.domain.pay.bankAccount.domain.entity.BankAccount;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
