+++ src/main/java/com/dangdang/server/domain/pay/connectionAccount/application/ConnectionAccountDatabaseService.java
@@ -1,19 +1,19 @@
-package com.dangdang.server.domain.connectionAccount.application;
+package com.dangdang.server.domain.pay.connectionAccount.application;
 
 import static com.dangdang.server.global.exception.ExceptionCode.BANK_ACCOUNT_INACTIVE;
 import static com.dangdang.server.global.exception.ExceptionCode.BANK_ACCOUNT_NOT_FOUND;
 import static com.dangdang.server.global.exception.ExceptionCode.PAY_MEMBER_NOT_FOUND;
 
-import com.dangdang.server.domain.bankAccount.domain.BankAccountRepository;
-import com.dangdang.server.domain.bankAccount.domain.entity.BankAccount;
-import com.dangdang.server.domain.bankAccount.exception.InactiveBankAccountException;
+import com.dangdang.server.domain.pay.bankAccount.domain.BankAccountRepository;
+import com.dangdang.server.domain.pay.bankAccount.domain.entity.BankAccount;
+import com.dangdang.server.domain.pay.bankAccount.exception.InactiveBankAccountException;
 import com.dangdang.server.domain.common.StatusType;
-import com.dangdang.server.domain.connectionAccount.domain.ConnectionAccountRepository;
-import com.dangdang.server.domain.connectionAccount.domain.entity.ConnectionAccount;
-import com.dangdang.server.domain.connectionAccount.dto.AddConnectionAccountRequest;
-import com.dangdang.server.domain.connectionAccount.exception.EmptyResultException;
-import com.dangdang.server.domain.payMember.domain.entity.PayMember;
-import com.dangdang.server.domain.payMember.domain.entity.PayMemberRepository;
+import com.dangdang.server.domain.pay.connectionAccount.domain.ConnectionAccountRepository;
+import com.dangdang.server.domain.pay.connectionAccount.domain.entity.ConnectionAccount;
+import com.dangdang.server.domain.pay.connectionAccount.dto.AddConnectionAccountRequest;
+import com.dangdang.server.domain.pay.connectionAccount.exception.EmptyResultException;
+import com.dangdang.server.domain.pay.payMember.domain.entity.PayMember;
+import com.dangdang.server.domain.pay.payMember.domain.entity.PayMemberRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -38,7 +38,7 @@ public class ConnectionAccountDatabaseService {
   @Transactional
   public ConnectionAccount addConnectionAccount(Long memberId,
       AddConnectionAccountRequest addConnectionAccountRequest) {
-    PayMember payMember = payMemberRepository.findByMember_Id(memberId)
+    PayMember payMember = payMemberRepository.findByMemberId(memberId)
         .orElseThrow(() -> new EmptyResultException(PAY_MEMBER_NOT_FOUND));
 
     BankAccount bankAccount = bankAccountRepository.findById(
