+++ src/main/java/com/dangdang/server/domain/pay/connectionAccount/domain/ConnectionAccountRepository.java
@@ -1,6 +1,6 @@
-package com.dangdang.server.domain.connectionAccount.domain;
+package com.dangdang.server.domain.pay.connectionAccount.domain;
 
-import com.dangdang.server.domain.connectionAccount.domain.entity.ConnectionAccount;
+import com.dangdang.server.domain.pay.connectionAccount.domain.entity.ConnectionAccount;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
