+++ src/main/java/com/dangdang/server/domain/pay/connectionAccount/domain/entity/ConnectionAccount.java
@@ -1,8 +1,8 @@
-package com.dangdang.server.domain.payUsageHistory.domain.entity;
+package com.dangdang.server.domain.pay.connectionAccount.domain.entity;
 
-import com.dangdang.server.domain.bankAccount.domain.entity.BankAccount;
+import com.dangdang.server.domain.pay.bankAccount.domain.entity.BankAccount;
 import com.dangdang.server.domain.common.BaseEntity;
-import com.dangdang.server.domain.payMember.domain.entity.PayMember;
+import com.dangdang.server.domain.pay.payMember.domain.entity.PayMember;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
@@ -12,23 +12,15 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import lombok.Getter;
-import org.hibernate.annotations.ColumnDefault;
 
 @Entity
 @Getter
-public class PayUsageHistory extends BaseEntity {
+public class ConnectionAccount extends BaseEntity {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Column(name = "pay_usage_history_id")
+  @Column(name = "connection_account_id")
   private Long id;
-
-  @Column(length = 255)
-  private String title;
-
-  @Column(columnDefinition = "INT UNSIGNED")
-  @ColumnDefault("0")
-  private Integer amount = 0;
 
   @ManyToOne(fetch = FetchType.LAZY)
   @JoinColumn(name = "pay_member_id")
@@ -38,6 +30,11 @@ public class PayUsageHistory extends BaseEntity {
   @JoinColumn(name = "bank_account_id")
   private BankAccount bankAccount;
 
-  protected PayUsageHistory() {
+  protected ConnectionAccount() {
+  }
+
+  public ConnectionAccount(PayMember payMember, BankAccount bankAccount) {
+    this.payMember = payMember;
+    this.bankAccount = bankAccount;
   }
 }
