+++ src/test/java/com/dangdang/server/domain/connectionAccount/application/ConnectionAccountDatabaseServiceTest.java
@@ -3,17 +3,18 @@ package com.dangdang.server.domain.connectionAccount.application;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-import com.dangdang.server.domain.bankAccount.domain.BankAccountRepository;
-import com.dangdang.server.domain.bankAccount.domain.entity.BankAccount;
-import com.dangdang.server.domain.bankAccount.exception.InactiveBankAccountException;
+import com.dangdang.server.domain.pay.bankAccount.domain.BankAccountRepository;
+import com.dangdang.server.domain.pay.bankAccount.domain.entity.BankAccount;
+import com.dangdang.server.domain.pay.bankAccount.exception.InactiveBankAccountException;
 import com.dangdang.server.domain.common.StatusType;
-import com.dangdang.server.domain.connectionAccount.domain.ConnectionAccountRepository;
-import com.dangdang.server.domain.connectionAccount.domain.entity.ConnectionAccount;
-import com.dangdang.server.domain.connectionAccount.dto.AddConnectionAccountRequest;
+import com.dangdang.server.domain.pay.connectionAccount.application.ConnectionAccountDatabaseService;
+import com.dangdang.server.domain.pay.connectionAccount.domain.ConnectionAccountRepository;
+import com.dangdang.server.domain.pay.connectionAccount.domain.entity.ConnectionAccount;
+import com.dangdang.server.domain.pay.connectionAccount.dto.AddConnectionAccountRequest;
 import com.dangdang.server.domain.member.domain.entity.Member;
 import com.dangdang.server.domain.member.domain.entity.MemberRepository;
-import com.dangdang.server.domain.payMember.domain.entity.PayMember;
-import com.dangdang.server.domain.payMember.domain.entity.PayMemberRepository;
+import com.dangdang.server.domain.pay.payMember.domain.entity.PayMember;
+import com.dangdang.server.domain.pay.payMember.domain.entity.PayMemberRepository;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
