+++ src/test/java/com/dangdang/server/domain/payMember/domain/entity/PayMemberRepositoryTest.java
@@ -2,6 +2,8 @@ package com.dangdang.server.domain.payMember.domain.entity;
 
 import com.dangdang.server.domain.member.domain.entity.Member;
 import com.dangdang.server.domain.member.domain.entity.MemberRepository;
+import com.dangdang.server.domain.pay.payMember.domain.entity.PayMember;
+import com.dangdang.server.domain.pay.payMember.domain.entity.PayMemberRepository;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -28,7 +30,7 @@ class PayMemberRepositoryTest {
     PayMember payMember = new PayMember(password, member);
     payMemberRepository.save(payMember);
 
-    PayMember findPayMember = payMemberRepository.findByMember_Id(member.getId()).get();
+    PayMember findPayMember = payMemberRepository.findByMemberId(member.getId()).get();
 
     Assertions.assertThat(findPayMember.getMember().getId()).isEqualTo(member.getId());
   }
